% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_write_cassandra_table.R
\name{cassandra_tbl}
\alias{cassandra_tbl}
\alias{spark_load_cassandra_table}
\alias{spark_write_cassandra_table}
\title{Load and write Cassandra tables}
\usage{
spark_load_cassandra_table(session, cass_keyspace, cass_tbl, spk_tbl_name,
  partition_filter = FALSE, select_cols = FALSE, cache_table = FALSE)

spark_write_cassandra_table(spk_tbl, cass_keyspace, cass_tbl)
}
\arguments{
\item{session}{The spark session}

\item{cass_keyspace}{A Cassandra keyspace}

\item{cass_tbl}{A Cassandra table/column family}

\item{spk_tbl_name}{The variable name of the loaded table in hive}

\item{partition_filter}{A server-side filter clause}

\item{select_cols}{A server-side select clause}

\item{spk_tbl}{A handle to a hive table to write}
}
\description{
Load and write Cassandra tables
}

