% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spark_get_master}
\alias{spark_get_master}
\title{spark_get_master}
\usage{
spark_get_master(port = 7077)
}
\arguments{
\item{port}{Spark master connection port}
}
\description{
Get the spark master connection string from the running host
}

