% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_get_session.R
\name{spark_get_session}
\alias{spark_get_session}
\title{spark_get_session}
\usage{
spark_get_session(sc)
}
\arguments{
\item{sc}{A spark context created with \code{sparklyr::spark_connect}}
}
\description{
Get the spark session
}

